refactor(web): move AuthProvider to auth folder
