+++ packages/web/src/pages/_app.tsx
@@ -28,7 +28,7 @@ import Cookies from 'cookies';
 import {theme} from '@ticoteco/ui';
 
 import buildRelayEnvironment, {AUTHORIZATION_KEY} from '../relay';
-import {AuthProvider} from '../providers/AuthProvider';
+import {AuthProvider} from '../auth/AuthProvider.js';
 
 export type TicoTecoAppProps = {
   readonly authorization: string;
